Шаблонизация лэйаута, главной страницы, блока проектов.
