+++ config/config.php
@@ -1,0 +1,6 @@
+<?php
+
+$config = [
+    'sitename' => 'Дела в порядке',  //название сайта
+    'timezone' => 'Europe/Moscow',  //временная зона
+];
