+++ config/db.php
@@ -1,0 +1,9 @@
+<?php
+
+$db = [
+    'host' => 'localhost',
+    'user' => 'root',
+    'password' => '',
+    'database' => '909803_doingsdone',
+    'timezone' => '+03:00',
+];
